Migrate fjs/bnf/descent to .f.mjs, split type-level API into types.ts
proof.f.ts stays TypeScript for now: it has a runtime dependency on
testlib.f.ts, still unmigrated. Updates the djs/tokenizer dependent
and stale module.f.ts references in bnf/ll1 and bnf's todo docs.
