+++ fjs/djs/tokenizer/proof.f.ts
@@ -1,4 +1,4 @@
-import { descentParser } from '../../bnf/descent/module.f.ts'
+import { descentParser } from '../../bnf/descent/module.f.mjs'
 import { stringToCodePointList, stringToList } from '../../text/utf16/module.f.mjs'
 import { toArray } from '../../types/list/module.f.mjs'
 import { jsGrammar, tokenizeString, descentParserCpOnly, tokenizeJs, tokenize } from './module.f.ts'
